Clean architecture

+++ models/phoneAuth.cs
@@ -4,6 +4,6 @@ namespace hiBuddy.models;
 
 public class phoneAuth
 {
-    public int phone_number { get; set; }
+    public long phone_number { get; set; }
     public String? pass { get; set; }
 }
